refactor(preset): code-split config & preset
